requirements.txt
articlemetaapi==1.26.5
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.2
hupper==1.6.1
idna==2.8
Jinja2==2.10
legendarium==2.0.2
lxml==4.3.1
Mako==1.0.8
MarkupSafe==1.1.1
more-itertools==6.0.0
nose==1.3.7
packtools==2.4
paginate==0.5.6
PasteDeploy==2.0.1
pathlib2==2.3.3
picles.plumber==0.11
plaster==1.0
plaster-pastedeploy==0.6
pluggy==0.9.0
ply==3.11
py==1.8.0
Pygments==2.3.1
pyramid==1.10.2
pyramid-debugtoolbar==4.5
pyramid-jinja2==2.8
pyramid-mako==1.0.2
pytest==4.3.1
pytest-cov==2.6.1
repoze.lru==0.7
requests==2.21.0
six==1.12.0
soupsieve==1.9
thriftpy==0.3.9
translationstring==1.3
urllib3==1.24.1
venusian==1.2.0
waitress==1.2.1
WebOb==1.8.5
WebTest==2.0.33
xylose==1.35.1
zope.deprecation==4.4.0
zope.interface==4.6.0
-e git://github.com/scieloorg/document-store.git@9422e6c8b8c25d5980c0c441b77293a42aaf77ec#egg=documentstore
SQLAlchemy==1.3.1
psycopg2-binary==2.7.7
